+++ skforecast/ForecasterAutoreg/tests/test_create_train_X_y.py
@@ -949,3 +949,5 @@ def test_create_train_X_y_output_when_window_features_and_exog_transformers_diff
     assert results[5] == expected[5]
     for k in results[6].keys():
         assert results[6][k] == expected[6][k]
+
+# TODO: tests with 2 window_features